requirements.txt
django
djangorestframework
django-crispy-forms
crispy-bootstrap5
pillow